requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
boto3==1.17.40
botocore==1.20.40
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
DateTime==4.3
idna==2.10
jmespath==0.10.0
multidict==5.1.0
numpy
pandas
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.3.6
schedule==1.0.0
six==1.15.0
slackclient==2.9.3
soupsieve==2.2.1
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
zope.interface==5.3.0
